requirements.txt
numpy==1.19.2
Pillow==8.0.1
streamlit==0.71.0
https://download.pytorch.org/whl/cpu/torch-1.3.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cu101/torchvision-0.4.2-cp36-cp36m-linux_x86_64.whl